Пофиксил ошибки с дао, сервисами, моделями

+++ src/main/java/models/Bill.java
@@ -24,7 +24,7 @@ public class Bill {
         this.id = id;
     }
 
-    public int getBonus() {
+    public double getBonus() {
         return bonus;
     }
